progress with sign in/sign out button
