+++ packages/editor/src/components/AuthWindows.tsx
@@ -1,12 +1,15 @@
+import {
+  createUserWithEmailAndPassword,
+  sendEmailVerification,
+  sendPasswordResetEmail,
+  signInWithEmailAndPassword,
+} from "firebase/auth";
 import { useState } from "react";
+import toast from "react-hot-toast";
 import { useRecoilState } from "recoil";
 import styled from "styled-components";
 import { AuthModalState, currentAuthModalState } from "../state/atoms.js";
-import {
-  logInUser,
-  loginWithGithub,
-  registerUser,
-} from "../utils/authUtils.js";
+import { authObject } from "../utils/authUtils.js";
 
 const MenuShadow = styled.div<{}>`
   position: absolute;
@@ -114,25 +117,80 @@ export const AuthMenuModal = () => {
 
   const [showForgotPass, setShowForgotPass] = useState(true);
 
-  const loginWrapper = () => {
-    console.log("hit 1");
-    console.log(logInUser());
-    if (logInUser()) {
-      console.log("hit 2");
-      closeModal();
+  const registerUser = () => {
+    var email = document.getElementById("reg_email") as HTMLInputElement;
+    var password = document.getElementById("reg_password") as HTMLInputElement;
+    var password_confirmation = document.getElementById(
+      "confirmed_password",
+    ) as HTMLInputElement;
+
+    if (email.value && password.value && password_confirmation.value) {
+      if (password.value != password_confirmation.value) {
+        toast.error("Password and confirmation must match!");
+        return;
+      }
+      createUserWithEmailAndPassword(authObject, email.value, password.value)
+        .then((userCredential) => {
+          // Signed up
+          sendEmailVerification(userCredential.user);
+          toast.success("Verification email sent, please check your email");
+          closeModal();
+        })
+        .catch((error) => {
+          const errorCode = error.code;
+          const errorMessage = error.message;
+          toast.error(errorMessage);
+        });
+    } else {
+      toast.error("Cannot register, field is blank");
     }
   };
 
-  const registerWrapper = () => {
-    if (registerUser()) {
-      closeModal();
+  const logInUser = () => {
+    var email = document.getElementById("login_email") as HTMLInputElement;
+    var password = document.getElementById(
+      "login_password",
+    ) as HTMLInputElement;
+
+    if (email.value && password.value) {
+      signInWithEmailAndPassword(authObject, email.value, password.value)
+        .then((userCredential) => {
+          // Signed up
+          toast.success("Logged in successfully!");
+          closeModal();
+        })
+        .catch((error) => {
+          const errorCode = error.code;
+          const errorMessage = error.message;
+          toast.error(errorMessage);
+        });
+    } else {
+      toast.error("Cannot log in, field is blank");
     }
   };
 
-  const forgotPassWrapper = () => {
-    if (registerUser()) {
-      setShowForgotPass(false);
+  const forgotPassword = () => {
+    var email = document.getElementById("login_email") as HTMLInputElement;
+    if (email.value) {
+      sendPasswordResetEmail(authObject, email.value)
+        .then(() => {
+          toast.success(`Sent password reset link to ${email.value}`);
+          setShowForgotPass(false);
+          // Password reset email sent!
+          // ..
+        })
+        .catch((error) => {
+          const errorCode = error.code;
+          const errorMessage = error.message;
+          toast.error(errorMessage);
+        });
+    } else {
+      toast.error("Please enter the account email above");
     }
+  };
+
+  const loginWithGithub = () => {
+    toast.error("Github OAuth not yet implemented");
   };
 
   return (
@@ -150,12 +208,12 @@ export const AuthMenuModal = () => {
                 placeholder="Password"
                 id="login_password"
               />
-              <FormButton onClick={loginWrapper}>Log In</FormButton>
+              <FormButton onClick={logInUser}>Log In</FormButton>
               <AdditionalOptions onClick={toggleMode}>
                 New user? Register
               </AdditionalOptions>
               {showForgotPass ? (
-                <AdditionalOptions onClick={forgotPassWrapper}>
+                <AdditionalOptions onClick={forgotPassword}>
                   Forgot Password?
                 </AdditionalOptions>
               ) : null}
@@ -176,7 +234,7 @@ export const AuthMenuModal = () => {
                 placeholder="Confirm Password"
                 id="confirmed_password"
               />
-              <FormButton onClick={registerWrapper}>Register</FormButton>
+              <FormButton onClick={registerUser}>Register</FormButton>
               <AdditionalOptions onClick={toggleMode}>
                 Return to Login
               </AdditionalOptions>
