+++ packages/editor/src/components/Settings.tsx
@@ -1,21 +1,41 @@
-import { useCallback } from "react";
+import { useEffect, useState } from "react";
 import { useRecoilState, useRecoilStateLoadable } from "recoil";
 import {
   AuthModalState,
   currentAuthModalState,
   settingsState,
 } from "../state/atoms.js";
+import { authObject, signOutWrapper } from "../utils/authUtils.js";
 import BlueButton from "./BlueButton.js";
 
 export default function Settings() {
   const [settings, setSettings] = useRecoilStateLoadable(settingsState);
-  const signOut = useCallback(() => {
-    setSettings((settings) => ({ ...settings, github: null }));
-  }, []);
+  // const signOut = useCallback(() => {
+  //   setSettings((settings) => ({ ...settings, github: null }));
+  // }, []);
 
   const [authModalState, setAuthModalState] = useRecoilState<AuthModalState>(
     currentAuthModalState,
   );
+
+  // idk if this actually works, would like to test more
+  // just need to fix the type err
+  const [loggedIn, setLoggedIn] = useState(null);
+
+  useEffect(() => {
+    const unsubscribe = authObject.onAuthStateChanged((user) => {
+      setLoggedIn(user);
+    });
+
+    return unsubscribe; // Clean up the observer on component unmount
+  }, []);
+
+  // useEffect(() => {
+  //   const currentUser = authObject.currentUser;
+  //   if (currentUser != null) {
+  //     setLoggedIn(true);
+  //   }
+  // }, []);
 
   const toggleLoginModal = () => {
     setAuthModalState({
@@ -56,20 +76,14 @@ export default function Settings() {
           />
         </label>
       </div>
-      {settings.contents.github === null && (
+      {loggedIn != null ? (
+        <div style={{ margin: "10px" }}>
+          {" "}
+          <BlueButton onClick={signOutWrapper}>sign out</BlueButton>
+        </div>
+      ) : (
         <div style={{ margin: "10px" }}>
           <BlueButton onClick={toggleLoginModal}>sign in</BlueButton>
-        </div>
-      )}
-      {settings.contents.github !== null && (
-        <div style={{ margin: "10px" }}>
-          <div>
-            signed into GitHub as{" "}
-            <span style={{ fontWeight: "bold" }}>
-              {settings.contents.github.username}
-            </span>
-          </div>
-          <BlueButton onClick={signOut}>sign out</BlueButton>
         </div>
       )}
     </div>
